Add day 8 part 1
I did not test this yet, because I don't have a quick way of installing
the containers and vector package on my system. This is ridiculous. The
Arch stack-static package fails with some obscure linker errors, and to
get a temporary ghc.withPackages shell from Nix I need to create a bunch
of files. I'll go with Python again next time then ... for now, I think
I can do with lists rather than Vector and IntMap. It will be
inefficient, but it will probably still work.