Start solving day 3 part 1 in Pony
I wanted to learn Pony for a long time. Advent of Code is a nice excuse.
I have no idea what I am doing, but so far I managed to read the bits
out of the input file, and I learned that I need to "consume" the line
to be able to iterate over its bytes.