Solve day 4 part 1 in Awk
I made the repo public and now it has almost 25% each for Awk, Haskell,
Python, and Pony, so for the lulz I need to get it balanced now. It's a
nice excuse to learn some more Awk beyond the basic oneliners that I can
write day to day. For example, I learned more about how arrays work in
Awk today, and how to do functions.