Simulate the monkey game
This is hacky in so many ways, but C is barebones. In Haskell I would
have written a mini-interpreter for the worry level expression language.
In C, we just abuse the fact that 0 does not occur as rhs in my input,
and that atoi returns it when it parses text without digits.