Solve 2022-22 part 2
It turns out the lay-out of the cube in my input is different from the
one in the example, after putting in the layout of my input, it works.

But this _really_ makes me think that there should be a more elegant
solution. Or maybe I should compute the neighbors from analyzing the
input or something. My current solution is bloody ugly. But hey, it
works ...