Add parser for 2022-23 input, in Pony
Let's give Pony one more try, it looks nice. And today's puzzle with
elves moving about sounds a lot like it could be a good fit for an actor
system, every elf an actor? (It could also be completely misleading and
maybe 1:1 elf:actor is the wrong approach, we'll see ...)