Almost solve day 4 part 2 in RCL
The resursive solution works, but is very inefficient. It computes the
right answer for the example input, but it takes a very long time to run
for the actual input. A better way would be to walk backwards through
the cards and compute their depth only once, but I don't have a feature
to reverse in RCL. So let's do it like this for now and, after a few
minutes it's already half-way and it should get faster near the end.