Speed up the RCL solution significantly
The forward pass that I did previously was recomputing the same things
all the time, which was possible but wasteful. It takes several minutes
to run. The backward pass solution that I implemented in this commit
takes only 60 milliseconds to run, which is pretty much instant to a
human pressing enter on a terminal; fast enough for me.

I wasn't sure this morning that I had the right machinery to do this,
but with the new std.range I don't even need to reverse lists or
anything. Slicing would still be nice to have, but with list
comprehensions, indexing, and std.range it is possible to get the same
result, so it's no longer urgent to add.