Implement the brute-force solution to part 2
It's not fast enough, I will have to use maths. If we joined each of the
parts by . instead of ?, then we could just do the number of
possibilities to the fifth power. But the ? does open up some options
that would not be included in that count. They would have to be options
with a # at the gap though, which should speed up the search. So maybe I
can break it down that way.