Correct a flaw in early-out, almost solve 12.2
It's still not *fast*, but it's fast enough that most inputs can
complete in a few minutes ... but then there are a few ones where the
entire input is just ??????????????????? and those still take ages!

Should I do a recursive solution then? Compute the number of ways to put
the groups on the slightly smaller template, and multiply by the
additional options that we get by growing the template by one?