Disable debug prints for day 12
It computes the right answer for the example, but not for my input
still. Also if I change the computation for the combinatorial x5 vs.
feeding in the expanded input (which is now very fast to compute), I get
a different result for my input (the latter is lower). For the example,
it doesn't matter. So something is still not right ...