12: Delete incorrect combinatorial method
It is not always true that if you join the maps with an . in between,
the number of ways to fill it out is just the product of the number of
ways to fill one out. A counterexample is "??#?? 1,1". For this pattern,
there are two ways to fill it out.

The pattern "??#?????#?? 1,1,1,1" has 8 ways to fill it out. 2 of those
have a # in the center, which leaves 6 to account for. If we do the
combinatorial combination, it assumes that the first two groups go in
the left "??#??" part, and the second two in the right part, but if we
join with ".", it is possible for three groups to go in the left or
right part: "#.#.#...#.." for example. So the combinatorial formula,
although it worked for the example, is wrong.

After some manual verification, the dynamic programming solution does
seem to produce the right values, but the answer it gives for my real
input is too low. It even gives a lower answer than the combinatorial
approach, which was already too low. But that's weird because in the
counterexample at this point, the combinatorial approach was under-
counting. Is the dynamic programming one under-counting different cases?