Port the Clojure solution to RCL
It doesn't work due to stack overflow. But maybe I can come up with a
clever way to avoid it? I could add a tail call optimization to the
evaluator, but that would be a lot of work and it's not useful to the
intended use case of RCL. But maybe there is a clever trick I can
express the computation in to create less deep stacks?