Convert day 18 into a "proper" F# project
I was running it with "dotnet fsi", in the REPL, but it was rather slow.
I thought it may be that the REPL is not jit-compiling stuff, so let me
make a "proper" project. (Fortunately I can avoid having an entire
"solution". Visual Studio was really nice 12 years ago but right now
it's overkill for what I need.)

Turns out, even when I do "dotnet run --configuration Release", it's
still slow. Back to the drawing board ...