Compute the frontier incrementally
Now we don't need O(result) storage, but only O(frontier), which is
O(sqrt(result)), since the result is an area but the frontier is its
boundary. Maybe this is sufficient to solve part 2, let's see.

For now I kept both to confirm that they compute the same result.