Sharpen the rules
I was doing this for the past two days anyway, and it doesn't make a
very big difference in what languages I can pick. In this case, it
eliminates Rust and Zig that would have been allowed if I want to stay
away from the top 50%. It should make it easier to push things out of
the disallowed set now :P I managed to push Rust out but now it's back
in. But that's okay, between OCaml and Clojure and C# it should be
possible to have a decent tool, and I can add stuff like Lua, Scala,
Idris, still.