Fail to converge for part 24.2
I have an idea for how to do it. It's silly, using gradient descent for
this, because everything is linear. But it works. When I run it on the
first three hailstones, it converges on an initial position. Then it has
*some* velocity in a well-defined position, but not yet at integer
coordinates. So I still need to find the right points on this line.

One way would be to compute all the intersections, then to find the
greatest common divisor of distances component-wise.