Update to latest version of RCL
This one has the new conditional syntax, and String.chars. It also has a
call depth limit that now unfortunately is too aggressive and breaks one
of the solutions that did work previously. But the proper solution to
this is not to bump the limit, but to implement tail calls.