Add masking optimization
This helps, we get much further now. Some are still slow (~seconds to
maybe a ~minute per machine), but so far at least when compiled in
release mode, it's working through them.

There are some problems though, where the mask is not really selective,
and it seems that it still gets stuck on those.

There is another step we can take though: if the button we are
inspecting is the *final* toggle for a light, we can also leverage that.
But that's for a follow-up commit.